% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.to.list.R, R/generic.helping.functions.R
\name{array.to.list}
\alias{array.to.list}
\title{A helping function to convert a 3d array into a list}
\usage{
array.to.list(.array)

array.to.list(.array)
}
\arguments{
\item{.array}{an array of dimensions axbxc to be converted into a list of length c with elements axb matrices}
}
\description{
A helping function to convert a 3d array into a list

helping functions
}
\examples{
array.to.list()
}
\keyword{data}
\keyword{preparing}
