% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{biasTerms}
\alias{biasTerms}
\title{Calculate bias-correction terms from full GWAS data}
\usage{
biasTerms(
  merged_data,
  dNames,
  harmonise_to = 1,
  pval_threshold = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{merged_data}{dataframe of merged exposure(s) and outcome GWAS}

\item{dNames}{length 3 list of with elements that are vectors of column names in merged_data corresponding to effect sizes, SEs, and effect alleles for all exposures and outcome. Order of elements in list vectors must correspond to the same phenotypes}

\item{harmonise_to}{phenotype to which effect alleles should be harmonised to}

\item{pval_threshold}{only SNPs with all P>pval_threshold will be used to calculate bias-correction terms}

\item{verbose}{TRUE/FALSE should some warnings be printed}
}
\description{
Calculate bias-correction terms from full GWAS data
}
\examples{
biasTerms()
}
