% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{causalPlot}
\alias{causalPlot}
\title{Plot causal estimates}
\usage{
causalPlot(
  fitList,
  exposure_names,
  qCI = 0.95,
  binaryOutcome = FALSE,
  colors = RColorBrewer::brewer.pal(length(fitList), "Spectral"),
  position_to_dodge = 0.8
)
}
\arguments{
\item{fitList}{list of fitted model objects using one of MRBEE.<x>()}

\item{exposure_names}{ordered names of exposures}

\item{qCI}{quantile for confidence intervals}

\item{binaryOutcome}{TRUE/FALSE: is outcome binary}

\item{colors}{colors for causal estimates made using different estimators}

\item{position_to_dodge}{0-1 dodging parameter for separation between causal estimates within exposures}
}
\description{
Plot causal estimates
}
\examples{
causalPlot()
}
