% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{checkBtTerms}
\alias{checkBtTerms}
\title{Check bias-correction terms can be calculated given data}
\usage{
checkBtTerms(merged_data, dNames)
}
\arguments{
\item{merged_data}{merged_data dataframe of merged exposure(s) and outcome GWAS}

\item{dNames}{length 3 list of with elements that are vectors of column names in merged_data corresponding to effect sizes, SEs, and effect alleles for all exposures and outcome. Order of elements in list vectors must correspond to the same phenotypes}
}
\description{
Check bias-correction terms can be calculated given data
}
\examples{
checkBtTerms()
}
