% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{classifyLoci}
\alias{classifyLoci}
\title{Classify outcome loci}
\usage{
classifyLoci(pleioP, exposureP, outcomeP, classifyingLociP = 5e-08)
}
\arguments{
\item{pleioP}{P-values for horizontal pleiotropy test for outcome}

\item{exposureP}{P-values for joint test of exposures directly from GWAS}

\item{outcomeP}{P-values for outcome directly from GWAS}

\item{classifyingLociP}{P-value to use in classifications of outcome loci into one of direct, mediation, pleiotropy (see published manuscript)}
}
\description{
Classify outcome loci
}
\examples{
classifyLoci()
}
