% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{EE.estimator.intercept}
\alias{EE.estimator.intercept}
\title{A function to perform non-iterative MRBEE with added intercept terms}
\usage{
EE.estimator.intercept(Ahat, Bhat, SigmaUU, SigmaUV, SigmaVV, Outliers = NULL)
}
\arguments{
\item{Ahat}{mxp matrix of standardized GWAS estimates for p exposures and m IVs}

\item{Bhat}{mxq matrix of standardized GWAS estimates for q outcomes and m IVs}

\item{SigmaUU}{pxpxm array of measurement error variance-covariance matrices for the p exposures and m IVs.}

\item{SigmaUV}{pxqxm array of measurement error covariance matrices for the p exposures, q outcomes, and m IVs.}

\item{SigmaVV}{qxqxm array of measurement error variance-covariance matrices for the q outcomes and m IVs.}

\item{Outliers}{vector of indices of items to remove from Ahat, Bhat, SigmaUU, SigmaUV, and SigmaVV from their m-index}
}
\description{
A function to perform non-iterative MRBEE with added intercept terms
}
\examples{
EE.estimator.intercept()
}
\keyword{MRBEE}
\keyword{intercept}
\keyword{with}
