% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EE.estimator.R
\name{EE.estimator}
\alias{EE.estimator}
\title{A function to perform non-iterative MRBEE}
\usage{
EE.estimator(
  A,
  B,
  SigmaUU,
  SigmaUV,
  SigmaVV,
  Outliers = NULL,
  perform.pleio = TRUE
)
}
\arguments{
\item{A}{mxp matrix of standardized GWAS estimates for p exposures and m IVs}

\item{B}{mxq matrix of standardized GWAS estimates for q outcomes and m IVs}

\item{SigmaUU}{pxpxm array of measurement error variance-covariance matrices for the p exposures and m IVs.}

\item{SigmaUV}{pxqxm array of measurement error covariance matrices for the p exposures, q outcomes, and m IVs.}

\item{SigmaVV}{qxqxm array of measurement error variance-covariance matrices for the q outcomes and m IVs.}

\item{Outliers}{vector of indices of items to remove from Ahat, Bhat, SigmaUU, SigmaUV, and SigmaVV from their m-index}

\item{perform.pleio}{logical indicating whether IV-specific horizontal pleiotropy tests should be performed using the statistic Spleio}
}
\description{
A function to perform non-iterative MRBEE
}
\examples{
EE.estimator()
}
\keyword{MRBEE}
