% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorCov.R
\name{errorCov}
\alias{errorCov}
\title{Estimate Error Covariance Matrix Using GWAS Insignificant Effects}
\usage{
errorCov(
  ZMatrix,
  Zscore.cutoff = 2,
  subsampling.ratio = 0.1,
  subsampling.time = 1000
)
}
\arguments{
\item{ZMatrix}{A matrix of Z-scores for exposure and outcome, with the outcome GWAS in the last column.}

\item{Zscore.cutoff}{The cutoff for significance. Defaults to 2.}

\item{subsampling.ratio}{The proportion of effects to subsample for each iteration. Defaults to 0.1.}

\item{subsampling.time}{The number of subsampling iterations. Defaults to 1000.}
}
\value{
A matrix representing the estimated error covariance.
}
\description{
This function estimates the error covariance matrix by subsampling a proportion of insignificant GWAS effects and calculating their correlation coefficients.
}
