% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.R
\name{ff}
\alias{ff}
\title{A helping function}
\usage{
ff(R, x)
}
\arguments{
\item{R}{any square matrix where row names match column names.}

\item{x}{matrix with column names to locate in R.}
}
\description{
This function finds the index locations of phenotypes in a correlation matrix.
}
\examples{
ind=sapply(colnames(X), function(h) ff(R,h))
ff()
}
\keyword{helping}
