% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fQ.R
\name{fQ}
\alias{fQ}
\title{A helping function to perform a global test of unbalanced horizontal pleiotropy}
\usage{
fQ(C, Est, Var)
}
\arguments{
\item{C}{a contrast matrix created by the iter.estimator() function (called from within the MRBEE() function) to extract intercept-relevant terms from a matrix}

\item{Est}{MRBEE causal estimates with first row corresponding to intercept terms}

\item{Var}{Variance-covariance matrix of MRBEE causal estimates}
}
\description{
A helping function to perform a global test of unbalanced horizontal pleiotropy
}
\examples{
fQ()
}
\keyword{contrast}
\keyword{testing}
