% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fS.R
\name{fS}
\alias{fS}
\title{A helping function to perform an IV-specific test of nonzero horizontal pleiotropy}
\usage{
fS(Bhati, Thetahat, Ahati, Upsilonhati)
}
\arguments{
\item{Bhati}{a qx1 vector (q outcomes) of standardized GWAS estimates for the ith SNP}

\item{Thetahat}{a pxq (q outcomes) matrix of causal effect estimates}

\item{Ahati}{a px1 vector (p exposures) of standardized GWAS estimates for the ith SNP}

\item{Upsilonhati}{an estimated variance-covariance calculated from within MRBEE(). Any consistent estimator of the variance of Bhati-t(Thetahat)\%*\%Ahati}
}
\description{
This function is called from within the MRBEE() function
}
\examples{
fS()
}
