% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{genomePleio}
\alias{genomePleio}
\title{Genome-wide horizontal pleiotropy testing}
\usage{
genomePleio(pD0, fit, n_divisions = 20, verbose = TRUE)
}
\arguments{
\item{pD0}{output from prepData() ran on full set of exposure and outcome merged GWAS summary statistics}

\item{fit}{object outputted from one of MRBEE.<x>()}

\item{n_divisions}{number of times to divide full set of GWAS summary statistics (for computational purposes)}

\item{verbose}{TRUE/FALSE should progress be printed}
}
\description{
Genome-wide horizontal pleiotropy testing
}
\examples{
genomePleio()
}
