% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{gwPleioPlot}
\alias{gwPleioPlot}
\title{Plotting results from genome-wide horizontal pleiotropy testing}
\usage{
gwPleioPlot(
  gwSp,
  leadSNPs,
  CHR_BP,
  GWsubInds = 1:length(gwSp$PleioStat),
  classifyingLociP = 5e-08,
  manhattanPUpper = 0.05
)
}
\arguments{
\item{gwSp}{object output of genomePleio()}

\item{leadSNPs}{rsIDs of lead outcome SNPs}

\item{CHR_BP}{vector of type "CHR_BP" positions corresponding to SNPs used in genome-wide horizontal pleiotropy testing}

\item{GWsubInds}{indices in full merged GWAS data to use in plotting (only change if you want plots for only a subset of all available SNPs)}

\item{classifyingLociP}{P-value to use in classifications of outcome loci into one of direct, mediation, pleiotropy (see published manuscript)}

\item{manhattanPUpper}{only SNPs with P<this threshold will be plotted in the Manhattan plot}
}
\description{
Plotting results from genome-wide horizontal pleiotropy testing
}
\examples{
gwPleioPlot()
}
