\name{hapmap3}
\alias{hapmap3}
\title{HapMap3 and UKBB Genotype SNP Data}
\description{
  A data frame containing combined SNP data from 1000 Genomes Project Phase 3 and UK Biobank (UKBB) genotypes. It includes a total of 1,664,852 SNPs.
}
\usage{
  data(hapmap3)
}
\format{
  A data frame with 1,664,852 rows and 3 variables:
    \describe{
      \item{SNP}{SNP identifier.}
      \item{A1}{Effect allele.}
      \item{A2}{Reference allele.}
    }
}
\source{
  1000 Genomes Project Phase 3 and UK Biobank genotype data.
}
\keyword{datasets}
