% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{imrp.mrbee.internal}
\alias{imrp.mrbee.internal}
\title{Internal MRBEE-IMRP function}
\usage{
imrp.mrbee.internal(
  by,
  bX,
  Rxx,
  rxy,
  max.iter = 15,
  max.eps = 0.001,
  pv.thres = 0.05,
  se.est = "robust",
  FDR = T,
  adjust.method = "Sidak",
  intercept = T
)
}
\description{
Internal MRBEE-IMRP function
}
\examples{
imrp.mrbee.internal()
}
