% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{imrpdetect}
\alias{imrpdetect}
\title{Helper}
\usage{
imrpdetect(
  x,
  theta,
  Rww,
  rwy,
  se.est = "ordinal",
  FDR = T,
  adjust.method = "Sidak"
)
}
\description{
Helper
}
\examples{
imrpdetect()
}
