% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter.estimator.R
\name{iter.estimator}
\alias{iter.estimator}
\title{A helping function to perform MR using MRBEE}
\usage{
iter.estimator(
  A,
  B,
  SigmaUU,
  SigmaUV,
  SigmaVV,
  PleioPThreshold,
  FDR = FALSE,
  FDR.alpha = 0.05,
  Outliers = NULL
)
}
\arguments{
\item{A}{mxp matrix of standardized exposure GWAS estimates}

\item{B}{mxq matrix of standardized exposure GWAS estimates}

\item{UU}{pxpxm array of measurement error variance-covariance matrices for the p exposures and m IVs.}

\item{UV}{pxqxm array of measurement error covariance matrices for the p exposures, q outcomes, and m IVs.}

\item{VV}{qxqxm array of measurement error variance-covariance matrices for the q outcomes and m IVs.}
}
\description{
MRBEE is intended to be performed using the MRBEE() function, which calls the function below.
}
\examples{
iter.estimator()
}
