% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE.BIC}
\alias{MRBEE.BIC}
\title{Find optimal tuning parameters for MRBEE methods using BIC}
\usage{
MRBEE.BIC(prepDataList, method = "IMRP")
}
\arguments{
\item{prepDataList}{the direct output of prepData(), a list of length 5.}

\item{method}{The specific MRBEE method, one of imrp, median, ipod, mix, or ml}
}
\description{
This function returns optimal tuning parameters for MRBEE methods using BIC
}
\examples{
MRBEE.BIC()
}
\keyword{MRBEE-BIC}
