% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE.IMRP.Egger.R
\name{MRBEE.IMRP.Egger}
\alias{MRBEE.IMRP.Egger}
\title{Estimate Causal Effect with MRBEE}
\usage{
MRBEE.IMRP.Egger(
  by,
  bX,
  byse,
  bXse,
  Rxy,
  max.iter = 30,
  max.eps = 1e-04,
  pv.thres = 0.05,
  var.est = "variance",
  FDR = TRUE,
  adjust.method = "Sidak",
  maxdiff = 3
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome.}

\item{bX}{A matrix (n x p) of the GWAS effect sizes of p exposures.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome.}

\item{bXse}{A matrix (n x p) of the GWAS effect size SEs of p exposures.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix of the p exposures and outcome. The last one should be the outcome.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to 30.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to 1e-4.}

\item{pv.thres}{P-value threshold in pleiotropy detection. Defaults to 0.05.}

\item{var.est}{Method for estimating the variance of residual in pleiotropy test. Can be "robust", "variance", or "ordinal". Defaults to "variance", which uses the sample variance of the residuals.}

\item{FDR}{Logical. Whether to apply the FDR to convert the p-value to q-value. Defaults to TRUE.}

\item{adjust.method}{Method for estimating q-value. Defaults to "Sidak".}

\item{maxdiff}{The maximum difference between the MRBEE causal estimate and the initial estimator. Defaults to 3.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy
}
\description{
This function estimates the causal effect using a bias-correction estimating equation with an intercept, considering potential pleiotropy and measurement errors.
}
