% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE.IMRP.UV.R
\name{MRBEE.IMRP.UV}
\alias{MRBEE.IMRP.UV}
\title{Univariable MR Estimation using MRBEE Method}
\usage{
MRBEE.IMRP.UV(
  by,
  bx,
  byse,
  bxse,
  Rxy,
  max.iter = 30,
  max.eps = 1e-04,
  pv.thres = 0.05,
  var.est = "variance",
  FDR = TRUE,
  adjust.method = "Sidak",
  var.method = "sandwich",
  sampling.time = 100
)
}
\arguments{
\item{by}{Vector of GWAS effect sizes for the outcome (n x 1).}

\item{bx}{Vector of GWAS effect sizes for the exposure (n x 1).}

\item{byse}{Vector of standard errors (SE) for the GWAS effect sizes of the outcome (n x 1).}

\item{bxse}{Vector of SE for the GWAS effect sizes of the exposure (n x 1).}

\item{Rxy}{Correlation matrix (p+1 x p+1) of the exposures and outcome, with the outcome being the last.}

\item{max.iter}{Maximum number of iterations for the estimation process. Defaults to 30.}

\item{max.eps}{Tolerance level for convergence. Defaults to 1e-4.}

\item{pv.thres}{P-value threshold for pleiotropy detection. Defaults to 0.05.}

\item{var.est}{Method for estimating the standard error in the pleiotropy test. Can be "robust", "variance", or "ordinal". Defaults to "variance", which uses the sample variance of the residuals.}

\item{FDR}{Logical indicating whether to apply False Discovery Rate (FDR) correction. Defaults to TRUE.}

\item{adjust.method}{Method for estimating q-values, defaults to "Sidak".}

\item{var.method}{Method for estimating variance of causal effect, defaults to "sandwich".}

\item{sampling.time}{Number of bootstrap resamples when var.method is not "sandwich". Defaults to 100.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy
}
\description{
This function performs univariable Mendelian Randomization (MR) analysis using
the MRBEE method.
}
