% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE.IPOD}
\alias{MRBEE.IPOD}
\title{MRBEE with IPOD adjustment}
\usage{
MRBEE.IPOD(
  pd,
  max.iter = 15,
  max.eps = 0.001,
  penalty = "SCAD",
  tau = 3,
  intercept = T
)
}
\arguments{
\item{pd}{output of prepData() on the IV set}

\item{max.iter}{maximum number of iterations to perform until convergence}

\item{max.eps}{tolerance (smaller values are stricter)}

\item{penalty}{type of penalty, one of "SCAD", "MCP", "lasso"}

\item{tau}{tuning parameter lambda for SCAD, MCP, and lasso penalties}

\item{intercept}{TRUE/FALSE should an intercept be included in causal estimation}
}
\description{
MRBEE with IPOD adjustment
}
\examples{
MRBEE.IPOD()
}
