% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE.Median}
\alias{MRBEE.Median}
\title{MRBEE with quantile estimation}
\usage{
MRBEE.Median(pd, tau = 0.5, h = 1, intercept = T, gradient = F)
}
\arguments{
\item{pd}{output of prepData() on the IV set}

\item{tau}{quantile to estimate (tau=0.5 is median regression)}

\item{h}{tuning parameter in (0,Inf), should be kept relatively large (> ~0.5). Computation is faster for larger values}

\item{intercept}{TRUE/FALSE should an intercept be included in causal estimation}

\item{gradient}{TRUE/FALSE should the Hessian of the loss function be used in optimization}
}
\description{
MRBEE with quantile estimation
}
\examples{
MRBEE.Median()
}
