% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE.Mix}
\alias{MRBEE.Mix}
\title{MRBEE assuming a mixture distribution for IVs}
\usage{
MRBEE.Mix(
  pd,
  max.iter = 20,
  max.eps = 1e-04,
  FDR = T,
  bic = T,
  outlier = F,
  thres = 0.05
)
}
\arguments{
\item{pd}{output of prepData() on the IV set}

\item{max.iter}{maximum number of iterations to perform until convergence}

\item{max.eps}{tolerance (smaller values are stricter)}

\item{FDR}{TRUE/FALSE should Benjamini-Hochberg FDR adjustment be used when identifying horizontally pleiotropic IVs}

\item{bic}{TRUE/FALSE should BIC be calculated}

\item{intercept1}{TRUE/FALSE should an intercept be included in causal estimation for the first of 2 IV sets}

\item{intercept2}{TRUE/FALSE should an intercept be included in causal estimation for the second of 2 IV sets}

\item{aic}{TRUE/FALSE should AIC be calculated}
}
\description{
MRBEE assuming a mixture distribution for IVs
}
\examples{
MRBEE.Mix()
}
