% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE.MLqe}
\alias{MRBEE.MLqe}
\title{MRBEE with likelihood-weighting of IVs}
\usage{
MRBEE.MLqe(
  pd,
  max.iter = 15,
  max.eps = 0.001,
  q = 0.95,
  method = "ordinal",
  intercept = T
)
}
\arguments{
\item{pd}{output of prepData() on the IV set}

\item{max.iter}{maximum number of iterations to perform until convergence}

\item{max.eps}{tolerance (smaller values are stricter)}

\item{q}{tuning parameter in (0,1)}

\item{method}{method to apply when constructing IV weights}

\item{intercept}{TRUE/FALSE should an intercept be included in causal estimation}
}
\description{
MRBEE with likelihood-weighting of IVs
}
\examples{
MRBEE.MLqe()
}
