% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE}
\alias{MRBEE}
\title{A function to perform MR using MRBEE}
\usage{
MRBEE(prepDataList, PleioPThreshold = 0.05, FDR = FALSE, FDR.alpha = 0.05)
}
\arguments{
\item{prepDataList}{the direct output of prepData(), a list of length 5.}

\item{PleioPThreshold}{The P-value threshold below which a specific IV will be considered horizontally pleiotropic and removed from causal estimation using pleio_test() (see paper).}

\item{FDR}{logical. If you want to use Benjamini-Hochberg FDR correction instead of defining a P-value threshold for determing horizontal pleiotropy evidence (as in the argument PleioPThreshold), set to TRUE.}

\item{FDR.alpha}{the uncorrected Type I error rate to be corrected by FDR.}
}
\description{
This function performs multivariable/multivariate MR using MRBEE.
}
\examples{
MRBEE()
}
\keyword{MRBEE}
