% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{pleio.test}
\alias{pleio.test}
\title{A function to perform an IV-specific test of nonzero horizontal pleiotropy for many IVs at once}
\usage{
pleio.test(
  Ahat,
  Bhat,
  Thetahat,
  CovThetahat,
  SigmaUU,
  SigmaUV,
  SigmaVV,
  Outliers = NULL
)
}
\arguments{
\item{Ahat}{a mxp matrix (p exposures) of standardized GWAS estimates for the m SNPs}

\item{Bhat}{a mxq matrix (q outcomes) of standardized GWAS estimates for the m SNPs}

\item{Thetahat}{a pxq matrix of causal effect estimates from MRBEE}

\item{CovThetahat}{An estimate of the variance-covariance matrix of Thetahat}

\item{SigmaUU}{an pxpxm array of variance-covariance matrices of measurement errors for the p exposures}

\item{SigmaUV}{an pxqxm array of covariance matrices of measurement errors for the p exposures and q outcomes}

\item{SigmaVV}{an qxqxm array of variance-covariance matrices of measurement errors for the q outcomes}

\item{Outliers}{vector of indices of items to remove from Ahat, Bhat, SigmaUU, SigmaUV, and SigmaVV from their m-index}
}
\description{
A function to perform an IV-specific test of nonzero horizontal pleiotropy for many IVs at once
}
\examples{
pleio.test()
}
\keyword{IMRP}
\keyword{SNPs}
\keyword{for}
\keyword{many}
\keyword{pleiotropy}
\keyword{test}
