% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{prepData}
\alias{prepData}
\title{A function to prepare GWAS data for MRBEE}
\usage{
prepData(bT, IVInds = 1:nrow(bT$EstHarm), outcome_index = 1, verbose = TRUE)
}
\arguments{
\item{betaX}{a mxp matrix of standardized GWAS estimates for p exposures and m genetic instrumental variables.}

\item{betaY}{a mxq matrix of standardized GWAS estimates for q outcomes and m genetic instrumental variables.}

\item{seX}{a mxp matrix of standardized GWAS standard errors for p exposures and m genetic instrumental variables (columns match order of betaX columns).}

\item{seY}{a mxp matrix of standardized GWAS standard errors for p outcomes and m genetic instrumental variables (columns match order of betaY columns).}

\item{R}{(p+q)x(p+q) matrix of correlations between GWAS estimates. See corrMatrix.py software.}
}
\description{
This function prepares GWAS data for MR with MRBEE.
}
\examples{
prepData()
}
\keyword{data}
\keyword{preparing}
