% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{residualPlot}
\alias{residualPlot}
\title{Plot residuals from MR}
\usage{
residualPlot(pd, fit, PleioPThreshold = 0.05)
}
\arguments{
\item{pd}{output of prepData() on the IV set}

\item{fit}{fitted model object using one of MRBEE.<x>()}

\item{PleioPThreshold}{P-value threshold for determining if a specific IV has sufficient evidence of horizontal pleiotropy}
}
\description{
Plot residuals from MR
}
\examples{
residualPlot()
}
