% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spleio.R
\name{Spleio}
\alias{Spleio}
\title{A wrapping function to perform an IV-specific test of nonzero horizontal pleiotropy for many IVs at once}
\usage{
Spleio(prepDataList, Thetahat, CovThetahat, Outliers = NULL)
}
\arguments{
\item{prepDataList}{the direct output of prepData(), a list of length 5.}

\item{Outliers}{vector of indices of IVs to remove}
}
\description{
A wrapping function to perform an IV-specific test of nonzero horizontal pleiotropy for many IVs at once
}
\examples{
Spleio()
}
\keyword{IMRP}
\keyword{horizontal}
\keyword{pleiotropy}
\keyword{test}
