% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssum.R
\name{ssum}
\alias{ssum}
\title{A helping function to add array elements of a list}
\usage{
ssum(.list)
}
\arguments{
\item{.list}{a list of three-dimensional arrays to be cumulatively summed}
}
\description{
A helping function to add array elements of a list
}
\examples{
ssum()
}
\keyword{data}
\keyword{preparing}
