% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.all.R
\name{subset.all}
\alias{subset.all}
\title{A helping function to subset the set of IVs}
\usage{
\method{subset}{all}(A, B, SigmaUU, SigmaUV, SigmaVV, Outliers)
}
\arguments{
\item{A}{mxp matrix of standardized GWAS estimates for p exposures and m IVs}

\item{B}{mxq matrix of standardized GWAS estimates for q outcomes and m IVs}

\item{SigmaUU}{pxpxm array of measurement error variance-covariance matrices for the p exposures and m IVs.}

\item{SigmaUV}{pxqxm array of measurement error covariance matrices for the p exposures, q outcomes, and m IVs.}

\item{SigmaVV}{qxqxm array of measurement error variance-covariance matrices for the q outcomes and m IVs.}

\item{Outliers}{vector of indices of items to remove from Ahat, Bhat, SigmaUU, SigmaUV, and SigmaVV from their m-index}
}
\description{
A helping function to subset the set of IVs
}
\examples{
subset.all()
}
\keyword{data}
\keyword{preparing}
