% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{weight.mrbee.internal}
\alias{weight.mrbee.internal}
\title{Internal weighting of MRBEE}
\usage{
weight.mrbee.internal(pd, weight = rep(1, length(by)), intercept = T)
}
\arguments{
\item{pd}{output of prepData() on the IV set}

\item{weight}{vector of weights to apply to IVs}

\item{intercept}{TRUE/FALSE should an intercept be included in causal estimation}
}
\description{
Internal weighting of MRBEE
}
\examples{
weight.mrbee.internal()
}
