% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfuns.R
\name{MRBEE.IMRP}
\alias{MRBEE.IMRP}
\title{MRBEE with IMRP adjustment}
\usage{
MRBEE.IMRP(
  prepDataList,
  PleioPThreshold = 0.05,
  FDR = FALSE,
  FDR.alpha = 0.05,
  inds = 1:nrow(prepDataList$betaX)
)
}
\arguments{
\item{prepDataList}{output of prepData() on the IV set}

\item{PleioPThreshold}{P-value threshold for determining if a specific IV has sufficient evidence of horizontal pleiotropy to be removed from causal estimation}

\item{FDR}{TRUE/FALSE should Benjamini-Hochberg FDR adjustment be used when identifying horizontally pleiotropic IVs}

\item{FDR.alpha}{Type I error rate for FDR adjustment}

\item{inds}{indices ot IV set to use - only change if you want to use an a-priori defined subset of all IVs}

\item{UU}{pxpxm array of measurement error variance-covariance matrices for the p exposures and m IVs.}

\item{UV}{pxqxm array of measurement error covariance matrices for the p exposures, q outcomes, and m IVs.}

\item{VV}{qxqxm array of measurement error variance-covariance matrices for the q outcomes and m IVs.}
}
\description{
MRBEE with IMRP adjustment
}
\examples{
MRBEE.IMRP()
}
